+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 v0.0.1-dev
 -------
 - Project Initialization
+- Support GET Method
+- Routing
+- Jinja2 Template Rendering